+++ packages/examples/src/langium/worker/statemachine-server.ts
@@ -1,0 +1,5 @@
+import { start } from './startLanguageServer.js';
+
+declare const self: DedicatedWorkerGlobalScope;
+
+start(self, 'statemachine-server');