+++ packages/examples/src/langium/wrapperStatemachine.ts
@@ -13,9 +13,24 @@ const startEditor = async () => {
         alert('Editor was already started!');
         return;
     }
-    const langiumGlobalConfig = await createLangiumGlobalConfig();
+
+    // init first worker regularly
+    const stateMachineWorkerRegular = loadStatemachinWorkerRegular();
+    const langiumGlobalConfig = await createLangiumGlobalConfig(stateMachineWorkerRegular);
     await wrapper.initAndStart(langiumGlobalConfig, document.getElementById('monaco-editor-root'));
-    const langiumGlobalConfig2 = await createLangiumGlobalConfig();
+
+    // init second worker with port for client and worker
+    const stateMachineWorkerPort = loadStatemachinWorkerPort();
+    // use callback to receive message back from worker independent of the message channel the LSP is using
+    stateMachineWorkerPort.onmessage = (event) => {
+        console.log('Received message from worker: ' + event.data);
+    };
+    const channel = new MessageChannel();
+    stateMachineWorkerPort.postMessage({
+        port: channel.port2
+    }, [channel.port2]);
+
+    const langiumGlobalConfig2 = await createLangiumGlobalConfig(stateMachineWorkerPort, channel.port1);
     await wrapper2.initAndStart(langiumGlobalConfig2, document.getElementById('monaco-editor-root2'));
 
     vscode.commands.getCommands().then((x) => {
@@ -42,13 +57,24 @@ export const run = async () => {
     }
 };
 
-export const loadStatemachinWorker = () => {
+export const loadStatemachinWorkerRegular = () => {
     // Language Server preparation
-    const workerUrl = new URL('./dist/worker/statemachineServerWorker.js', window.location.href);
+    const workerUrl = new URL('./src/langium/worker/statemachine-server.ts', window.location.href);
     console.log(`Langium worker URL: ${workerUrl}`);
 
     return new Worker(workerUrl, {
         type: 'module',
-        name: 'Statemachine LS',
+        name: 'Statemachine Server Regular',
+    });
+};
+
+export const loadStatemachinWorkerPort = () => {
+    // Language Server preparation
+    const workerUrl = new URL('./src/langium/worker/statemachine-server-port.ts', window.location.href);
+    console.log(`Langium worker URL: ${workerUrl}`);
+
+    return new Worker(workerUrl, {
+        type: 'module',
+        name: 'Statemachine Server Port',
     });
 };